# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Eazy Talks / Match Vibe Backend API
  description: |
    Auto-generated from Express route files (scripts/generate-openapi.mjs).
    Base path for API modules is /api/v1. Regenerate after route changes.
  version: 1.0.0
  contact:
    name: zztherapy backend
servers:
  - url: http://localhost:3000
    description: Local development
  - url: https://api.yourdomain.com
    description: Production (set PUBLIC_API_BASE_URL host)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT or Firebase ID token
      description: Staff JWT from /auth/admin-login, /agency-login, /bd-login OR Firebase ID token for mobile
    webhookSignature:
      type: apiKey
      in: header
      name: X-Signature
      description: Provider HMAC signature on raw body (Stream, Razorpay, Cloudflare)
    metricsToken:
      type: apiKey
      in: header
      name: X-Metrics-Token
      description: Optional ops token when METRICS_TOKEN env is set
  schemas:
    SuccessEnvelope:
      type: object
      properties:
        success:
          type: boolean
          example: true
        data:
          type: object
...
paths:
...
  "/api/v1/chat/webhook":
    post:
      tags:
        - Chat
      operationId: post__api_v1_chat_webhook
      summary: Chat POST /api/v1/chat/webhook
      security:
        - webhookSignature: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '4XX':
          description: Client error
        '5XX':
          description: Server error
...
  "/api/v1/payment/webhook":
    post:
      tags:
        - Payment
      operationId: post__api_v1_payment_webhook
      summary: Payment POST /api/v1/payment/webhook
      security:
        - webhookSignature: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '4XX':
          description: Client error
        '5XX':
          description: Server error
...
  "/api/v1/stream/webhook":
    post:
      tags:
        - Stream
      operationId: post__api_v1_stream_webhook
      summary: Stream POST /api/v1/stream/webhook
      security:
        - webhookSignature: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '4XX':
          description: Client error
        '5XX':
          description: Server error
...
  "/api/v1/video/webhook":
    post:
      tags:
        - Video
      operationId: post__api_v1_video_webhook
      summary: Video POST /api/v1/video/webhook
      security:
        - webhookSignature: []
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessEnvelope'
        '4XX':
          description: Client error
        '5XX':
          description: Server error
...